requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
Pillow==7.2.0
promise==2.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.11
wrapt==1.12.1
zipp==3.4.0
